Add a tsconfig.json to examples/tools directory.
Allow using Typescript to write command line examples.

We need a project file to let VSCode IntelliSense find the mupdf module.
We need to set the module resolution scheme to nodenext so that it can
pick up the mupdf module, and we should not pick up the root tsconfig
which is intended for the distribution library.