This adds a dedicated "Search" button to the UI for simple viewer.
Addresses issues #78 & #98
A listener for search input doesn't work well
with language input methods which convert
words into kanji or Chinese characters.
This adds a dedicated search button and
ensures that the previous & next buttons
are shown and hidden as required.
It also prints out better syntax for the search
information.