Update example tests.
Document that examples/tests runs with locally built mupdf. Most the scripts
import from ../../dist/mupdf so the README was not being truthful. Some scripts
import from npm, so fix those to run locally to be consistent.

Update package.json to NOT install a remote "mupdf" package, as that is not
necessary and may introduce errors where we are not testing what we think
we are testing.

Use TextEncoder to convert strings to ArrayBuffer for passing around.

Use mupdf objects directly from mupdf package (not via re-export).

Clean up authentication test.

Clean up crypting test to create a valid PDF file with contents that will be encrypted.